+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes starting with v0.1.34 to this project will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# v0.1.55 (2022-01-29
+# v0.1.56 (2022-02-07)
+- **changed:** increase default bundle size to 100 MB to reduce number of tx and use `chunk/` endpoint instead of `tx/` endpoint used for bundles 10 MB or less.
+- **changed:** post transactions in chunks when uploading files greater than 10 MB to allow individual files of unlimited size to be uploaded.
+- **added:** include [upload_files](examples/upload_files.rs) example.
+
+# v0.1.55 (2022-01-29)
 - **added:** include `write-metaplex-items` in `upload-nfts` command.
 
 # v0.1.54 (2022-01-18)